Add @highlight.js library to highlight raw xml data
Github project @highlight.js is a syntax highlighter written in JavaScript.
We only need a couple of its files, but since git can't include only part of a project, this library is not linked to highlight.js but just copies the necessary files into dmarcts. We will have to keep an eye for any major updates.